Update: fix path

+++ shinyaigeek.dev/scripts/deleteDistAssets.ts
@@ -2,9 +2,11 @@ import path from "path";
 import fs from "fs";
 
 const deleteDistAssets = () => {
-  fs.rmdirSync(path.join(__dirname, "../../public"), {
-      recursive: true
-  });
+  if (fs.existsSync(path.join(__dirname, "../../public"))) {
+    fs.rmdirSync(path.join(__dirname, "../../public"), {
+      recursive: true,
+    });
+  }
 };
 
 deleteDistAssets();
